+++ enclone_vars/src/vars
@@ -873,7 +873,7 @@ notes:
 code:
 ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 name:     cred
-inputs:   ?
+inputs:   lvar_vdj
 limits:
 class:    lvar
 level:    cell-exact
@@ -883,7 +883,42 @@ brief:    median credibility assessed using GEX data
 page:     enclone help lvars
 avail:    public
 notes:
-code:
+code:     let mut credsx = Vec::<f64>::new();
+          for l in 0..ex.clones.len() {
+              let bc = &ex.clones[l][0].barcode;
+              let li = ex.clones[l][0].dataset_index;
+              if gex_info.pca[li].contains_key(&bc.clone()) {
+                  let mut creds = 0;
+                  let mut z = Vec::<(f64, String)>::new();
+                  let x = &gex_info.pca[li][&bc.clone()];
+                  for y in gex_info.pca[li].iter() {
+                      let mut dist2 = 0.0;
+                      for m in 0..x.len() {
+                          dist2 += (y.1[m] - x[m]) * (y.1[m] - x[m]);
+                      }
+                      z.push((dist2, y.0.clone()));
+                  }
+                  z.sort_by(|a, b| a.partial_cmp(b).unwrap());
+                  let top = n_vdj_gex[li];
+                  for i in 0..top {
+                      if bin_member(&vdj_cells[li], &z[i].1) {
+                          creds += 1;
+                      }
+                  }
+                  let pc = 100.0 * creds as f64 / top as f64;
+                  credsx.push(pc);
+              } else {
+                  credsx.push(0.0);
+              }
+          }
+          let credsx_unsorted = credsx.clone();
+          credsx.sort_by(|a, b| a.partial_cmp(b).unwrap());
+          let mut r = Vec::<String>::new();
+          for j in 0..credsx_unsorted.len() {
+              r.push(format!("{:.1}", credsx_unsorted[j]));
+          }
+          cell: r
+          exact: format!("{:.1}", median_f64(&credsx))
 ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
 name:     cred_cell
 inputs:   ?
